bugfix - support guarded union alias narrowing (#741)
