+++ src/backend/ir/scanners/mod.rs
@@ -1,13 +1,15 @@
 //! Feature scanners and collectors for IR codegen.
 //!
-//! This module centralizes feature detection logic. The functions here are pure analyzers over the parsed AST and do
-//! not mutate global state.
+//! This module centralizes feature detection logic. The functions here are pure analyzers over parsed AST or lowered
+//! IR and do not mutate global state.
 
+mod binding_usage;
 mod decorators;
 mod rust_crates;
 mod serde;
 mod this;
 
+pub(crate) use binding_usage::{binding_use_scan, expr_uses_binding_name};
 pub use rust_crates::collect_rust_crates;
 pub use serde::{detect_serde_non_import_usage, detect_serde_usage};
 pub use this::check_for_this_import;
